Remove raise layer from bottom left mod keys
It seems quite plausible that the Cmd/Opt being stuck is due to layer mods being left behind when raise is pressed, i.e. press cmd or opt first, then press raise, then release the cmd or opt. In that case the cmd or opt in the qwerty layer won't detect the release, so they're stuck on until next press/release cycle.

We can also try to mitigate this with the PREVENT_STUCK_MODIFIERS feature, but it's worth more thought on whether we want these critical mods to be overlaid in regular operation or not, since that means the sequence of Cmd-Raise-number vs Raise-Cmd-number would matter.